requirements.txt
# GridMind Sentinel — Dashboard Deployment Requirements
# Python 3.11 required.
# For Streamlit Community Cloud deployment (dashboard only — no Ollama/FastAPI needed).

# --- Agent Orchestration ---
langgraph>=0.2.0,<0.3.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0

# --- Vector Store ---
faiss-cpu>=1.8.0,<1.9.0

# --- Embeddings & Reranker ---
sentence-transformers>=3.0.0,<4.0.0

# --- BM25 Keyword Search ---
rank-bm25>=0.2.2,<0.3.0

# --- Signal Processing ---
PyWavelets>=1.6.0,<1.8.0

# --- Power Simulation ---
numpy>=1.26.0
scipy>=1.12.0

# --- Backend API (needed for src.* imports in evaluation module) ---
fastapi>=0.115.0,<0.116.0
uvicorn>=0.30.0

# --- Database ---
sqlalchemy>=2.0.0,<3.0.0

# --- Dashboard ---
streamlit>=1.40.0,<1.42.0
streamlit-autorefresh>=1.0.1

# --- Data Validation ---
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# --- Visualization ---
plotly>=5.0.0,<6.0.0
pandas>=2.0.0,<3.0.0

# --- Testing ---
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0

# --- Utilities ---
python-dotenv>=1.0.0
httpx>=0.27.0

# --- Long-term Memory ---
chromadb>=0.5.0,<0.6.0
